% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoid-package.R
\docType{package}
\name{nanoid}
\alias{nanoid}
\alias{nanoid-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/nanoid}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/nanoid/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
