% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{data}
\name{nanoid_dict}
\alias{nanoid_dict}
\title{Built in NanoID alternate character alphabet subsets}
\format{
An object of class \code{list} of length 9.
}
\usage{
nanoid_dict
}
\description{
The default alphabet subset is "\code{url_friendly}", consisting of upper/lower letters, numbers and
\verb{_}, \code{-}.
}
\details{
Other subsets include:
\itemize{
\item \code{numbers} = "\code{0123456789}"
\item \code{lowercase} = "\code{abcdefghijklmnopqrstuvwxyz}"
\item \code{uppercase} = "\code{ABCDEFGHIJKLMNOPQRSTUVWXYZ}"
\item \code{hexadecimal_lowercase} = "\verb{0123456789abcdef}"
\item \code{hexadecimal_uppercase} = "\verb{0123456789ABCDEF}"
\item \code{nolookalikes} = "\verb{346789ABCDEFGHJKLMNPQRTUVWXYabcdefghijkmnpqrtwxyz}"
\item \code{nolookalikes_safe} = "\verb{6789BCDFGHJKLMNPQRTWbcdfghjkmnpqrtwz}"
\item \code{alphanumeric} = "\verb{0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ}"
}
}
\keyword{datasets}
