% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{NANOIDgenerate}
\alias{NANOIDgenerate}
\title{Generate a NanoID}
\usage{
NANOIDgenerate(size = 21L, alphabet = nanoid_dict[["url_friendly"]])
}
\arguments{
\item{size}{(int) number of characters in the generated NanoID. Defaults to 21. Larger values
decrease the chance of a collision}

\item{alphabet}{character subset to use in the NanoID generation process. By default,
\code{\link[=NANOIDgenerate]{NANOIDgenerate()}} uses a "URL friendly" character set. See \link{nanoid_dict} for other,
built-in alphabet subsets.}
}
\description{
Generate a NanoID
}
\examples{
NANOIDgenerate() # default to a URL friendly character set and char length 21
NANOIDgenerate(10) # same as ^^ but char length of 10
NANOIDgenerate(10, nanoid_dict[["nolookalikes"]]) # numbers and English alphabet without lookalike
}
