% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoid-package.R
\docType{package}
\name{nanoid}
\alias{nanoid}
\alias{nanoid-package}
\title{Tools to Generate Short, Secure, URL-friendly, and Unique String Identifiers}
\description{
Nano IDs are comparable to UUID v4 (random-based). They have a similar number of random
bits in the ID (126 in Nano ID and 122 in UUID), so they have similar collision probabilities.
Nano ID also use a bigger alphabet, so a similar number of random bits are packed in just 21
symbols instead of 36. The C++ cooe implements the algorithm specified by Andrey Sitnik.
}
\references{
\preformatted{- NanoID algorithm and JavaScript implementation <https://github.com/ai/nanoid>
- C++ implementation: <https://github.com/mcmikecreations/nanoid_cpp>
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/nanoid}
  \item Report bugs at \url{https://github.com/hrbrmstr/nanoid/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
